--- GetUserDataProcessingActivity/function.json
@@ -1,10 +1,0 @@
-{
-  "bindings": [
-    {
-      "name": "name",
-      "type": "activityTrigger",
-      "direction": "in"
-    }
-  ],
-  "scriptFile": "../dist/GetUserDataProcessingActivity/index.js"
-}
